+++ .gitignore
@@ -1,3 +1,4 @@
+babytalk_corpus.txt
 data/
 *.pt
 *.pth
